+++ app/forms/candidate_interface/degree_wizard.rb
@@ -228,12 +228,9 @@ module CandidateInterface
       if existing_degree.present?
         existing_degree.update(attributes_for_persistence)
         clear_state!
-      else
-        # Avoid jumping steps
-        unless attributes_for_persistence.slice(:qualification_type, :institution_name, :subject, :grade, :start_year, :award_year).values.any?(&:blank?)
-          ApplicationQualification.create!(attributes_for_persistence)
-          clear_state!
-        end
+      elsif all_attributes_for_persistence_present?
+        ApplicationQualification.create!(attributes_for_persistence)
+        clear_state!
       end
     end
 
@@ -437,6 +434,12 @@ module CandidateInterface
         NO => NOT_APPLICABLE,
         I_DO_NOT_KNOW => UNKNOWN,
       }[grade]
+    end
+
+    def all_attributes_for_persistence_present?
+      attributes_for_persistence.slice(
+        :qualification_type, :institution_name, :subject, :grade, :start_year, :award_year
+      ).values.all?(&:present?)
     end
 
     def self.map_completed(application_qualification)
